+++ src/print.js
@@ -1,27 +1,15 @@
 const { breakParent, concat, hardline, lineSuffix } = require("prettier").doc.builders;
+const { printComments } = require("./utils");
 const nodes = require("./nodes");
 
 module.exports = (path, opts, print) => {
-  const { type, body, comment, start } = path.getValue();
+  const { type, body, comments, start } = path.getValue();
 
   if (type in nodes) {
     const printed = nodes[type](path, opts, print);
 
-    if (comment) {
-      if (comment.start < start) {
-        return concat([
-          comment.break ? breakParent : "",
-          comment.body,
-          hardline,
-          printed
-        ]);
-      }
-
-      return concat([
-        printed,
-        comment.break ? breakParent : "",
-        lineSuffix(` ${comment.body}`)
-      ]);
+    if (comments) {
+      return printComments(printed, start, comments);
     }
     return printed;
   }